Fix `ffmpeg:uarchive` stalled at the end
In previous version, we use named pipe and fifo. However it will cause ffmpeg to miss the "close", aka EOF, if ffmpeg is busy and wait too long. The comment in code shows the log. And, this is why we use shell redirection `-f xxx - > ./fifo.mkv` instead of `-f xxx ./fifo.mkv` in the old shell script implement.